requirements.txt
asgiref==3.3.0
boto3==1.16.58
botocore==1.19.58
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3.1
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.10.0
django-js-asset==1.2.2
django-storages==1.11.1
django-tempus-dominus==5.1.2.13
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
urllib3==1.25.11
